Move sizeOfDirectory() method
